fix error

+++ examples/transformers/led/generate.py
@@ -15,7 +15,7 @@ TXT = "My friends are <mask> but they eat too many carbs."
 # load from modelscope (run `modelscope download --model allenai/led-base-16384` ahead)
 checkpoint_path = "/home/user/.cache/modelscope/hub/models/allenai/led-base-16384"
 tokenizer = AutoTokenizer.from_pretrained(checkpoint_path)
-model = LEDForConditionalGeneration.from_pretrained(checkpoint_path, mindspore_dtype=ms.float16)
+model = LEDForConditionalGeneration.from_pretrained(checkpoint_path, mindspore_dtype=ms.float32)
 
 
 input_ids = Tensor(tokenizer([TXT], return_tensors="np")["input_ids"])
